Handle Invoice Data Passed into Editor (Fix New Invoice Creation)
Instead of rendering an id (if we dont already have one) before passing into the editor component, we instead generate an ID within the component. This fixes the issue where we can not save away new invoices that were not retrieved from already existing stored invoices - as the generated ID kept changing as we passed it in.